pruneman

+++ .bash_aliases
@@ -6,6 +6,16 @@ alias grep="rg"
 
 if [ -x /usr/bin/podman ]; then
 pruneman() {
+    local y
+    echo
+    echo "this will prune all podman extraneous objects"
+    echo "make sure any containers of interest ARE RUNNING"
+    echo
+    read -p "continue? (Y/n)" y
+    echo
+    if [[ "$y" == "n" ]]; then
+        return
+    fi
     podman image prune
     podman container prune
     podman volume prune
